Release: version 0.0.12

+++ src/Action/HashRecords.php
@@ -20,7 +20,7 @@ class HashRecords
         $this->hashColumn = $hashColumn;
     }
 
-    public function execute(array $exclude = [])
+    public function execute(array $include = [], array $exclude = [])
     {
         // Resolve model via morph map if available
         $modelClass = Relation::getMorphedModel($this->modelClass) ?? $this->modelClass;
@@ -30,8 +30,8 @@ class HashRecords
             Log::error("Prevent Same Shit: Invalid model class => {$modelClass} or enforce morph map not implemented, terminating");
             throw new Exception("Invalid model class => {$modelClass} or enforce morph map not implemented, terminating");
         }
-        
-        $model = new $modelClass;        
+
+        $model = new $modelClass;
         $table = $model->getTable();
         $columns = Schema::getColumnListing($table);
         if(!in_array($this->hashColumn, $columns)){
@@ -39,22 +39,35 @@ class HashRecords
             throw new Exception("Specified hashed column not found, terminating");
             return false;
         }
+        $hasSoftDeletes = (in_array("deleted_at", $columns))? true:false;
 
         $excludedFromHash = ['id',$hashColumn,'created_at','updated_at'];
         foreach ($exclude as $column) {
             array_push($excludedFromHash, $column);
         }
 
+        $includedInHash = $include;
+
+        echo("ℹ️ Prevent Same Shit: Included columns\n");
+        print_r($includedInHash);
+        Log::info("Prevent Same Shit: Included columns", $includedInHash);
+
         echo("ℹ️ Prevent Same Shit: Excluded columns\n");
         print_r($excludedFromHash);
         Log::info("Prevent Same Shit: Excluded columns", $excludedFromHash);
 
-        $modelClass::chunk(100, function ($models) use ($hashColumn, $excludedFromHash) {
+        $modelClass::chunk(100, function ($models) use ($hashColumn, $includedInHash , $excludedFromHash, $hasSoftDeletes) {
             foreach ($models as $model) {
                 $data = $model->attributesToArray();
+
+                if (!empty($includedInHash)) {
+                    $data = array_intersect_key($data, array_flip($includedInHash));
+                }
+
                 foreach ($excludedFromHash as $column) {
                     unset($data[$column]);
                 }
+
                 ksort($data);
                 $payload = json_encode($data);
 
@@ -70,9 +83,11 @@ class HashRecords
                 }catch(Exception $e){
                     Log::warning("Prevent Same Shit: Duplicate row detected at id: {$model->id}");
                     echo("⚠️ Prevent Same Shit: Duplicate row detected at id: {$model->id}\n");
-                    $model->delete();
-                    echo("ℹ️ Prevent Same Shit: Duplicate row deleted\n");
-                    Log::info('Prevent Same Shit: Duplicate row deleted');
+                    if($hasSoftDeletes){
+                        $model->delete();
+                        echo("ℹ️ Prevent Same Shit: Duplicate row soft deleted\n");
+                        Log::info('Prevent Same Shit: Duplicate row soft deleted');
+                    }
                 }
             }
         });
